pyproject.toml
[project]
name = "mindloop"
version = "0.2.0"
description = "Recurrent Transformer Language Model - CS5804 Group Project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=3.5.0",
    "matplotlib>=3.10.1",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
    "transformers>=4.51.3",
    "tqdm>=4.66.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
cpu = ["torch>=2.6.0", "torchvision>=0.21.0"]
cu124 = ["torch>=2.6.0", "torchvision>=0.21.0"]

[tool.uv]
conflicts = [[{ extra = "cpu" }, { extra = "cu124" }]]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu124", extra = "cu124" },
]
torchvision = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu124", extra = "cu124" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
